more tests

+++ server/controllers/reviewController/reviewController.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
 const mapUserToReview = require('../../middleware/mapUserToReview');
 const createReviewValidation = require('../../middleware/createReviewValidation');
+const updateReviewValidation = require('../../middleware/updateReviewValidation');
+const deleteReviewValidation = require('../../middleware/deleteReviewValidation');
 const ReviewRepository = require('./reviewRepository');
 const { validationResult } = require('express-validator');
 const { noReviews, serverError } = require('../../constants/httpResponses');
@@ -22,7 +24,6 @@ router.post('/:space', mapUserToReview, createReviewValidation, async (req, res)
     try {
         const validation = validationResult(req);
         const errorsPresent = !validation.isEmpty();
-
         if (errorsPresent) return res.status(400).json({ error: validation.array() });
 
         const newReview = await repo.create(req.body);
@@ -32,9 +33,13 @@ router.post('/:space', mapUserToReview, createReviewValidation, async (req, res)
     }
 });
 
-router.put('/:space', mapUserToReview, async (req, res) => {
+router.put('/:space', mapUserToReview, updateReviewValidation, async (req, res) => {
     const repo = new ReviewRepository(req.params.space);
     try {
+        const validation = validationResult(req);
+        const errorsPresent = !validation.isEmpty();
+        if (errorsPresent) return res.status(400).json({ error: validation.array() });
+
         await repo.update(req.body);
         return res.sendStatus(204);
     } catch (error) {
@@ -42,9 +47,13 @@ router.put('/:space', mapUserToReview, async (req, res) => {
     }
 });
 
-router.delete('/:space', mapUserToReview, async (req, res) => {
+router.delete('/:space', mapUserToReview, deleteReviewValidation, async (req, res) => {
     const repo = new ReviewRepository(req.params.space);
     try {
+        const validation = validationResult(req);
+        const errorsPresent = !validation.isEmpty();
+        if (errorsPresent) return res.status(400).json({ error: validation.array() });
+
         await repo.delete(req.body);
         return res.sendStatus(204);
     } catch (error) {
